Refactored to use byte/tinyint throughout. Added ability to link players to games.

+++ DungeonsAndDragons-DB/Tables/Enums/Abilities.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[Abilities]
 (
-    AbilityId INT IDENTITY(1,1) 
+    AbilityId TINYINT IDENTITY(1,1) 
 		CONSTRAINT Abilities_AbilityId_PK PRIMARY KEY,
     Name VARCHAR(50) NOT NULL,
 	FlavourText VARCHAR(MAX)
