+++ DungeonsAndDragons-DB/Tables/Enums/PowerSources.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[PowerSources]
 (
-    PowerSourceId INT IDENTITY(1,1) 
+    PowerSourceId TINYINT IDENTITY(1,1) 
 		CONSTRAINT PowerSources_PowerSourceId_PK PRIMARY KEY,
     Name VARCHAR(50) NOT NULL,
 	FlavourText VARCHAR(MAX)
